op-acceptance-tests: remove INITIAL_L2_BLOCK variable, and fetch latest block number from EL
